+++ ideadensity/__init__.py
@@ -1,4 +1,5 @@
 from .idea_density_rater import cpidr
 from .depid import depid
+from .utils.version_utils import reproducibility_info
 
-__all__ = ['cpidr', 'depid']
+__all__ = ['cpidr', 'depid', 'reproducibility_info']